feat: Refactor model structure and associations; replace dynamic imports with explicit model imports
