--- src/config/config.json
@@ -1,24 +1,0 @@
-{
-  "development": {
-    "username": "root",
-    "password": "root",
-    "database": "teaestate",
-    "host": "localhost",
-    "port":3306,
-    "dialect": "mysql"
-  },
-  "test": {
-    "username": "root",
-    "password": "root",
-    "database": "database_test",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
-  },
-  "production": {
-    "username": "root",
-    "password": "root",
-    "database": "database_production",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
-  }
-}
