+++ src/services/calendarService.js
@@ -1,4 +1,4 @@
-import { Calendar } from "../models";
+import { Calendar } from "../../models";
 
 export const getAllCalendars = async () => {
     return await Calendar.findAll();
