+++ src/services/teaWeightService.js
@@ -1,6 +1,6 @@
 // src/services/teaWeightService.js
 
-import { TeaWeight } from '../models';
+import { TeaWeight } from '../../models';
 
 export const getAllTeaWeights = async () => {
     return await TeaWeight.findAll();
